+++ great_expectations/cli/docs.py
@@ -1,13 +1,16 @@
 import os
-import sys
 
 import click
 
-from great_expectations import DataContext
-from great_expectations import exceptions as ge_exceptions
 from great_expectations.cli.cli_logging import logger
-from great_expectations.cli.util import cli_message, cli_message_list
-from great_expectations.core.usage_statistics.usage_statistics import send_usage_message
+from great_expectations.cli.util import (
+    cli_message,
+    cli_message_list,
+    load_data_context_with_error_handling,
+)
+from great_expectations.core.usage_statistics.usage_statistics import (
+    send_usage_message,
+)
 
 
 @click.group()
@@ -34,41 +37,10 @@ def docs():
     default=True,
 )
 def docs_build(directory, site_name, view=True):
-    """
-    Build Data Docs for a project.
-
-    :param directory:
-    :param site_name: name for the first-level keys in the "data_docs_sites" section of the configuration
-    :param view:
-    """
-
-    context = None
-    try:
-        failed = True
-        context = DataContext(directory)
-        build_docs(context, site_name=site_name, view=view)
-        failed = False
-        send_usage_message(
-            data_context=context,
-            event="cli.docs.build",
-            success=True
-        )
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        sys.exit(1)
-    except ge_exceptions.PluginModuleNotFoundError as err:
-        cli_message(err.cli_colored_message)
-        sys.exit(1)
-    except ge_exceptions.PluginClassNotFoundError as err:
-        cli_message(err.cli_colored_message)
-        sys.exit(1)
-    finally:
-        if failed and context is not None:
-            send_usage_message(
-                data_context=context,
-                event="cli.docs.build",
-                success=False
-            )
+    """ Build Data Docs for a project."""
+    context = load_data_context_with_error_handling(directory)
+    build_docs(context, site_name=site_name, view=view)
+    send_usage_message(data_context=context, event="cli.docs.build", success=True)
 
 
 @docs.command(name="list")
@@ -80,47 +52,30 @@ def docs_build(directory, site_name, view=True):
 )
 def docs_list(directory):
     """List known Data Docs Sites."""
-    context = None
-    try:
-        failed = True
-        context = DataContext(directory)
-        docs_sites_url_dicts = context.get_docs_sites_urls()
-        docs_sites_strings = [
-            " - <cyan>{}</cyan>: {}".format(docs_site_dict["site_name"], docs_site_dict["site_url"])\
-            for docs_site_dict in docs_sites_url_dicts
-        ]
+    context = load_data_context_with_error_handling(directory)
 
-        if len(docs_sites_strings) == 0:
-            cli_message("No Data Docs sites found")
-            failed = False
-            send_usage_message(
-                data_context=context,
-                event="cli.docs.list",
-                success=True
-            )
-            return
+    docs_sites_url_dicts = context.get_docs_sites_urls()
+    docs_sites_strings = [
+        " - <cyan>{}</cyan>: {}".format(docs_site_dict["site_name"], docs_site_dict["site_url"])\
+        for docs_site_dict in docs_sites_url_dicts
+    ]
 
-        if len(docs_sites_strings) == 1:
-            list_intro_string = "1 Data Docs site found:"
-
-        if len(docs_sites_strings) > 1:
-            list_intro_string = "{} Data Docs sites found:".format(len(docs_sites_strings))
+    if len(docs_sites_strings) == 0:
+        cli_message("No Data Docs sites found")
+    else:
+        list_intro_string = _build_intro_string(docs_sites_strings)
         cli_message_list(docs_sites_strings, list_intro_string)
-        failed = False
-        send_usage_message(
-            data_context=context,
-            event="cli.docs.list",
-            success=True
-        )
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-    finally:
-        if failed and context is not None:
-            send_usage_message(
-                data_context=context,
-                event="cli.docs.list",
-                success=False
-            )
+
+    send_usage_message(data_context=context, event="cli.docs.list", success=True)
+
+
+def _build_intro_string(docs_sites_strings):
+    doc_string_count = len(docs_sites_strings)
+    if doc_string_count == 1:
+        list_intro_string = "1 Data Docs site found:"
+    elif doc_string_count > 1:
+        list_intro_string = f"{doc_string_count} Data Docs sites found:"
+    return list_intro_string
 
 
 def build_docs(context, site_name=None, view=True):