+++ great_expectations/cli/init.py
@@ -7,8 +7,6 @@ from great_expectations import DataContext
 from great_expectations import exceptions as ge_exceptions
 from great_expectations.cli.datasource import \
     add_datasource as add_datasource_impl
-from great_expectations.cli.toolkit import \
-    create_expectation_suite as create_expectation_suite_impl
 from great_expectations.cli.docs import build_docs
 from great_expectations.cli.init_messages import (
     BUILD_DOCS_PROMPT,
@@ -24,6 +22,8 @@ from great_expectations.cli.init_messages import (
     SLACK_SETUP_PROMPT,
     SLACK_WEBHOOK_PROMPT,
 )
+from great_expectations.cli.toolkit import \
+    create_expectation_suite as create_expectation_suite_impl
 from great_expectations.cli.util import cli_message, is_sane_slack_webhook
 from great_expectations.exceptions import (
     DataContextError,
@@ -167,9 +167,3 @@ def _slack_setup(context):
 
 def _get_full_path_to_ge_dir(target_directory):
     return os.path.abspath(os.path.join(target_directory, DataContext.GE_DIR))
-
-
-def _complete_onboarding(target_dir):
-    DataContext.create(target_dir)
-    cli_message(ONBOARDING_COMPLETE)
-    return True